Address PR: move serializable namespace into schema
