+++ src/primitive.ts
@@ -55,40 +55,3 @@ export type JsonScalar = null | boolean | number | string;
 export type JsonValue = NestedValue<JsonScalar>;
 export type JsonObject = NestedObject<JsonScalar>;
 export type JsonArray = NestedArray<JsonScalar>;
-
-/**
- * Lists of types which are JSON serializable
- */
-export namespace Serializeable {
-  /**
-   * JSON serializable type of NodeReference
-   */
-  export interface NodeReference {
-    id: JsonScalar;
-    path: JsonScalar[];
-  }
-
-  /**
-   * JSON serializable type of NodeSnapshot in GraphSnapshot
-   *
-   * This is used when doing extract and restore cache's stage
-   */
-  export interface GraphSnapshot {
-    [key: string]: {
-      nodeSnapshotType: Serializeable.NodeSnapshotType,
-      inbound?: Serializeable.NodeReference[],
-      outbound?: Serializeable.NodeReference[],
-      data?: JsonValue,
-    };
-  }
-
-  /**
-   * Type of NodeSnapshot. We need this so that when deserialize, we can
-   * correct create NodeSnapshot.
-   *
-   */
-  export const enum NodeSnapshotType {
-    EntitySnapshot,
-    ParameterizedValueSnapshot,
-  }
-}
